move proto file
