+++ grpc_node/proto/.gitignore
@@ -1,0 +1,2 @@
+#protoc build file
+*.ts
